test: Add e2e test for import image-bundle (#192)
This test uses Docker to spin up a container based on a KinD image which
contains containerd. `mindthegap` binary and a created image bundle are
copied to the container, `import image-bundle` is run, and then the image
ID is retrieved from containerd via `crictl`. A known image ID is used to
ensure that it does not change and to prevent regressions.