fix: Allow TLS to pull chart from OCI repository (#804)
This fix allows us to test the `serve bundle` command using a TLS
client.

Previously, we created our own registryClient. We were responsible for
configuring TLS for the client, but did not. We now allow helm to create
the client for us. Because we previously did not customize the client in
any way, we only gain functionality, and lose none.

---------

Co-authored-by: Jimmi Dyson <jimmidyson@gmail.com>