fix: Reload registry TLS certificate when it is updated on the file system (#805)
Watch the filesystem for changes to the certificate, and reload when it
is updated.

Introduces a dependency on
https://pkg.go.dev/sigs.k8s.io/controller-runtime/pkg/certwatcher. This
does not appear to impact the executable size.

The first commit is a test. On its own, it fails. With the fix in the
second commit, the test passes.

Fixes #803

(Note: This is stacked on #804)

---------

Co-authored-by: Jimmi Dyson <jimmidyson@gmail.com>