feat: Use ArchiveFS instead of extracting where possible (#845)
This removes the need for significant disk space when pushing or
serving large bundles.

`mindthegap` creates a registry via the defacto standard
`github.com/distribution/distribution` module. Previously this registry
would always be configured to serve from a local directory. When
creating bundles (via `mindthegap create bundle`) the pulled data is
written to a directory and then can be archived to a tar that can then
be used by the other capablities of `mindthegap` (i.e. `push`, `import`,
and `serve`) - all of which only require read-only access to the data.

Previously in these read-only modes, the `mindthegap` bundle - a `tar`
file - was extracted to a temporary directory and then the internal
registry was configured to use that directory as the root directory of
the `filestorage`. This meant that for large bundles (e.g. large
air-gapped bundles) that significant disk space was required to extract
the bundle. The bundle that `mindthegap` creates is not compressed so
the disk space required is equal to the size of the bundle btw so a 10GB
bundle requires an extra 10GB of space for extraction prior to being
able to run the main functionality of the `push`/`import`/`serve`
command. This implementation detail is hidden to the user.

This PR introduces a new storage type (implementing the
[`StorageDriver`](https://github.com/distribution/distribution/blob/main/registry/storage/driver/storagedriver.go#L560)
interface from the Docker distribution project) that uses a `fs.FS`
implementation built on top of an archive (`ArchiveFS`) so that the
internal registry in `mindthegap` does not need to perform the
extraction and therefore saves on the temporary disk space, as well as
being faster as there is no initial extraction required.

All write operations are unimplemented by this storage driver as the
archive is read-only and that satisfies the use cases in `mindthegap`.