feat: Introduce merge-with-{overwrite,retain} push modes (#882)
These modes merge the pushed indexes with indexes already present in the
registry, while either retaining existing matching images in the index
or overwriting them. Matching is done on a platform basis, so if the
existing index includes `linux/amd64` image and a new one is pushed
then depending on the selected mode the existing image in the index
will be overwritten or retained.