fix: Correctly check not found errors in registry storage (#918)
This PR includes 2 commits, the first adding an e2e test that fails
before the fix from the second commit is applied.

Details of the fix:

Previously, this was using `os.IsNotExist` error checking, but since the
move to using `ArchiveFS` from `github.com/mholt/archives` this should
have been changed to `errors.Is(err, fs.ErrNotExist)` which is now the
standard way to check filesystem errors with `fs.FS` based filesystems.